[#24127,#23140,#24245] Docdb: Reload relcache/catcache upon setting yb_read_time
Summary:
We encountered several cloning issues when we cloned to a time before some DDL operations occurred. Two examples are:

- Cloning to a time before renaming a column that is used by an index #24127.
- Cloning to a time before an index is dropped, if tablet splitting takes place #23140.
- Cloning to a time before a rename of a constraint #24245.

The main reason for the previous error is that the `relcache/catcache` of the PostgreSQL backend has the latest catalog version cached for some entries, even after setting the `yb_read_time` to an earlier time. This means that the PostgreSQL backend returns the latest table schemas for some cached relations, while the uncached relations are returned as of `yb_read_time`. This leads to inconsistent YSQL schemas generated by `ysql_dump`. For example, in the first case above, the generated `CREATE TABLE` command has the column name as of `yb_read_time` (e.g., `CREATE TABLE test_table (key INT PRIMARY KEY, value TEXT);` notice the column name `value`), while the `CREATE INDEX` command has the column name as of the latest point in time `value2` (e.g., `CREATE INDEX NONCONCURRENTLY test_table_idx ON public.test_table USING lsm (value2 HASH) SPLIT INTO 2 TABLETS;`).

The fix addresses this issue by resetting the catalog version in the pg backend upon setting (or unsetting) the `yb_read_time`. This leads to reloading the `relcache/catcache` which is populated with the catalog information as of the new `yb_read_time`, ensuring that all catalog data returned using this pg backend will be as of `yb_read_time`.

However, after setting the `yb_read_time` and clearing the caches, we encountered the following error when trying to execute any query:

```
error: query failed: ERROR:  The catalog snapshot used for this transaction has been invalidated: expected: 5, got: 4: MISMATCHED_SCHEMA
```

This occurs because the tserver is refusing to process the read requests from this pg backend, as it has an older catalog version ("4", as of the read time) compared to the latest catalog version ("5").

The catalog version check is useful as a protection mechanism for time-traveling queries. That is, we don't want to allow users to read the data as of a point in time before a ddl operation. We might allow time-traveling queries across DDLs in the future after enough testing is done on DDLs. Therefore, we want the following invariant to be true all the time: `The catalog version of a pg backend corresponds to the read time in this backend`.

To make reloading the caches at a specific time possible, a new GUC variable, `yb_disable_catalog_version_check`, has been added. This variable skips setting the catalog version in all the read requests from the current pg backend. This prevents the tserver from checking whether the read requests coming from this pg backend have the latest catalog version. This GUC is `false` by default and is only set to `true` in pg sessions created by `ysql_dump` when `read-time` is set. The purpose of this variable is to ignore the catalog version for sessions reading data from a past time, as they might legitimately have an older catalog version.

It is worth mentioning that we decided to introduce the variable `yb_disable_catalog_version_check` instead of disabling the catalog version check whenever `yb_read_time` is set. We don't want to allow users to query user data from a time when the catalog version is older than the latest version. Currently, `ysql_dump` is the only client that is allowed to read data as of an older catalog version, as it only sends queries to the pg catalog located in the master. This variable can be set to `true` in other scenarios once enough testing has been done to guarantee that consistent user data is returned even across DDLs.
Jira: DB-13020, DB-12082, DB-13141

Test Plan: ./yb_build.sh --cxx-test yb-admin-snapshot-schedule-test --gtest_filter ColocationAndRestoreType/YbAdminSnapshotScheduleTestWithYsqlColocationRestoreParam.PgsqlRenameColumnWithIndex/3

Reviewers: kfranz, myang, asrivastava, mihnea, dmitry, hsunder

Reviewed By: myang, dmitry

Subscribers: hsunder, mihnea, ybase, slingam, yql, jason

Differential Revision: https://phorge.dev.yugabyte.com/D40501