[PLAT-16479] Make single connection for YSQL upgrade the default
Summary:
Based on the most recent tests for YSQL upgrade it seems that using
a single connection for it makes the best use of resources for almost
negligible increase in the overall time. Therefore we are making
the single connection as the default. Here are the results for YSQL
upgrade with 105 DBs in a universe :-

========================================================
Mode                  |   Memory Increase |    Time    |
========================================================
Multiple connections  |     0.53 GB       |   1m 9s    |
Single connection     |       0           |   1m 28s   |
========================================================

Please see more details here :-
https://yugabyte.atlassian.net/issues/PLAT-16479

Test Plan: Ran UTs and Itests

Reviewers: anijhawan, daniel, nsingh

Reviewed By: nsingh

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41318