[BACKPORT 2024.2][PLAT-16473] Delete the 'Remove table' action from xCluster replication table list UI
Summary:
This diff deletes the 'Remove table' action from the xCluster replication table list UI.
This legacy remove table action has been replaced by the database & table selection modal in both
xCluster replication and xCluster disaster recovery user workflows.

Please also note that post-v2024.2.0, the `tables` field in an xCluster object no longer contains only the
tables in the xCluster config. It may include other related tables which belong to the same database
as a table which is in the xCluster config. For any api clients which need a list of tables current in the
xCluster config, you may filter the `tableDetails` field and omit tables which have the following statuses:
- ExtraTableOnSource
- ExtraTableOnTarget
- DroppedFromSource
- DroppedFromTarget
Using the remove table action from the xCluster tables list UI will send additional table ids if any table with
one of the above statuses exists in the list. Please use the select databases and tables action from the xCluster
actions dropdown to perform any addition or removal of tables to this issue. The select databases and tables modal
filters out all tables which are not part of the current xCluster config and not part of the user selection prior to
submitting.
For any branch/release containing this change (v2024.2.2 and beyond) the per-row 'Remove table' legacy action will no
longer be available. Going fowards, the only place users should go to for database/table additions/removals is the
select databases and tables modal.

Original commit: 24f936651d35e7f11d9e44960c63d44371154870 / D41321

Test Plan:
Create an xCluster config and add/remove tables as a sanity check.
The changes in this diff only remove a dropdown from the YBA UI and the corresponding modal.
The YBA xCluster UI should function as expected with only the per-row 'Remove table' action deleted.

{F326600}

Reviewers: rmadhavan, hzare, cwang, vbansal, kkannan, lsangappa

Reviewed By: cwang

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41524