[#25687] DocDB: Remove key value storage callbacks from vector LSM
Summary:
In initial design vector LSM was responsible for binding vector id with indexed table key.
So we added a way to store this binding in regular DB.
Currently vector id is generated externally to vector LSM, so it could avoid dealing with indexed table keys, and this logic could be implemented in higher level.
Jira: DB-14945

Test Plan: Jenkins

Reviewers: arybochkin

Reviewed By: arybochkin

Subscribers: yql, ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D41337