[PLAT-16517][PLAT-15743]: Update expression pushdown gflag in non-restart upgrade fashion.
Summary:
Added changes to manage the expression_pushdown gflag during upgrades, rollbacks, and finalization in a non-restart upgrade fashion.

YBA is required to configure expression_pushdown for YSQL major version upgrades due to database limitations. This flag must be set to disable expression_pushdown before the upgrade begins and remain disabled throughout the process until either a rollback is completed or finalization is triggered. Additionally, the flag should remain disabled for other universe tasks when the system is in a pre-finalize state.

If the user has enabled the flag, YBA will reject the upgrade request with a "Bad Request" response and prompt the user to disable the flag before proceeding.

This update also includes changes to restrict upgrades to versions starting from 2024.2.1+ builds.

Test Plan:
Manually tested the YSQL major version upgrade process on a universe by performing upgrade, rollback, and finalization. Verified that the expression_pushdown flag is set at the beginning of the upgrade and remains applied until either the rollback is completed or the upgrade is finalized.

Confirmed that the user receives a "Bad Request" response when attempting to upgrade a universe running a version lower than 2024.2.1.0-b1. Additionally, validated that the upgrade proceeds successfully when the universe version is greater than 2024.2.1.0-b1.

Reviewers: #yba-api-review, hsunder, sanketh, svarshney, anijhawan, sneelakantan

Reviewed By: #yba-api-review, sanketh

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41352