[PLAT-16211] [PLAT-16218] [PLAT-16222] [PLAT-16215] Add support for creating CipherTrust KMS config
Summary:
This diff addresses the following tickets:
1. [PLAT-16211] Added support for creating a CipherTrust KMS config.
2. [PLAT-16218] Add support for refresh token auth flow for CipherTrust KMS.
3. [PLAT-16222] Add validation for CipherTrust KMS.
4. [PLAT-16215] Hide the CipherTrust KMS feature behind a runtime config.

Runtime config: `yb.kms.allow_ciphertrust`.

Sample request:
```
curl --location 'http://localhost:9000/api/v1/customers/<CUSTOMER_UUID>/kms_configs/CIPHERTRUST' \
--header 'Content-Type: application/json' \
--header 'X-AUTH-YW-API-TOKEN: <AUTH_TOKEN>' \
--data '{
    "name": "ct1",
    "CIPHERTRUST_MANAGER_URL": "<CT_MANAGER_URL>",
    "REFRESH_TOKEN": "<REFRESH_TOKEN>",
    "KEY_ALGORITHM": "AES",
    "KEY_SIZE": 256,
    "KEY_NAME": "testkey4"
}'
```

Case 1:
If a key with the name `testkey4` already exists on the ciphertrust manager, we just validate it and reuse the key in the KMS config. We do a test encrypt and decrypt for a small random text.

Case 2:
If a key with the given name doesn't already exist on the ciphertrust manager, we will create it on the ciphertrust manager and use that in the KMS config.

For both the above cases, we will update the local auth config object with the correct details as on the ciphertrust manager.

Test Plan:
Manually tested multiple flows:
1. Create KMS config without pre-existing key on CT manager console.
2. Create KMS config with pre-existing key on CT manager console.
3. Create a KMS config with pre-existing key on CT manager console, but with different key algorithm and key size in the API. (Ensured the correct details are updated locally)

Will add UTs in a separate ticket.
Run UTs.
Run itests.

Reviewers: vkumar

Reviewed By: vkumar

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41358