[PLAT-15973] Add bootstrap summary to create xCluster Disaster Recovery (DR) modal
Summary:
A bootstrap summary (referred to as "full copy creation plan" in the YBA UI) page is used as part of the logic for determining whether we need to collect bootstrap params from the user.
As DR config creation always collects bootstrap params in the user flow from YBA UI, we did not include the bootstrap summary page here prior to this diff.

The motivation for adding this summary to the DR creation modal comes from wanting to remove any confusion about whether the DR creation processes will involve bootstrapping. Bootstrapping is a backup and restore operation which may take quite some time and though we indicate in the creation modal that tables with data will be bootstrapped, adding the bootstrap summary UI here removes any doubt regarding which tables will or will not be bootstrapped.

Test Plan:
Attempt xCluster disaster recovery config creation using a set of tables which require bootstrapping.
Verify the bootstrap summary page appears as expected.

Attempt xCluster disaster recovery config creation using a set of tables which do not require bootstrapping.
Verify the bootstrap summary page appears as expected and that the backup storage config field is
still required for form submission.
{F326610}

Reviewers: hzare, cwang, vbansal, rmadhavan, kkannan, lsangappa

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41359