[PLAT-16528] Generate swagger with array elements one per line
Summary:
The phab diff for API changes are very hard to make sense of. For example, when adding a new task type, there is a single large line that shows up as a changed line in the phab diff. This makes is impossible to review the exact task type that was added.
It would be good to have a better formatting of the diff, perhaps one item on each line.

Test Plan: sbt swaggerGen

Reviewers: anijhawan, #yba-api-review

Reviewed By: anijhawan, #yba-api-review

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41361