[#22612] YSQL, QueryDiagnostics: Add Dynamic Background Worker Registration and Termination
Summary:
Before this diff:
A background worker was started at the start of the server, and it stays switched on even if no bundles are being processed.

After this diff:
A background worker is started only if there is any bundle being diagnosed. And if all bundles are done processing, then the background worker is terminated.
Jira: DB-11522

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestYbQueryDiagnostics#testDynamicBackgroundWorker'

Reviewers: asaha, telgersma

Reviewed By: telgersma

Subscribers: svc_phabricator, telgersma, yql

Differential Revision: https://phorge.dev.yugabyte.com/D40430