[#14895, #25521] YSQL: Fix column mapping for partitioned table for INSERT ... ON CONFLICT DO UPDATE queries
Summary:
## The problem
Postgres allows a table partition to define its columns in an order that is different from the parent (root) table. Consider the following example where the base table defines columns in the order [a, b, c], while the partition defines columns in the order [c, b, a]:

```
CREATE TABLE base (a INT PRIMARY KEY, b TEXT, c FLOAT) PARTITION BY LIST (a);

-- Create a partition
CREATE TABLE part1 (c FLOAT, b TEXT, a INT NOT NULL);
ALTER TABLE base ATTACH PARTITION part1 FOR VALUES IN (1);
```

Further, a query may insert/update a tuple in a partition while referencing the parent table as follows:

```
INSERT INTO p  VALUES (1, 'bar') ON CONFLICT (a) DO UPDATE SET a=1;
```

This implies that the input tuple is received in the layout of the parent table, and needs to be converted into the format of the partition before insert/update.

Vanilla postgres largely takes care of this by maintaining a tuple conversion map, which maintains mapping between columns in the parent to that in the partition. This conversion is applied whenever the tuple is routed, and the tuple is inserted/updated in the correct format and table.

However, in the case of Yugabyte’s implementation of INSERT ON CONFLICT .. DO UPDATE, we rely on a list of modified columns that is passed down from the planner to prepare the update payload. There are cases where the destination partition is not resolved at planning time, which results in the parent table’s columns to be passed down which is then incorrectly used to apply the update on the child partition.

This revision explicitly maps the columns of the root to that of the partition whose tuple is being updated when:
 - The operation being performed is INSERT ... ON CONFLICT DO UPDATE
 - The partitioned table defines its columns in an order that is different from that of its partition root.

Further, this revision also enables vanilla PG's INSERT ... ON CONFLICT tests that were previously commented out because of the above bugs.
Jira: DB-4172, DB-14778

Test Plan:
Run the following tests:
```
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressPgDml#schedule'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressPartitions#yb_partitions_tests'
```

TBD: More tests will be added to yb_partitions_out_of_order.

Reviewers: aagrawal, amartsinchyk

Reviewed By: aagrawal

Subscribers: steve.varnau, smishra, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D41110