[PLAT-14303][dr] Support rollback of switchover operation
Summary:
This diff adds the rollback api `POST   /customers/:cUUID/tasks/:tUUID/rollback` with empty body. Then it reads the task params associated with task uuid `:tUUID` from the YBA database and creates a task to roll it back.

It also adds the rollback button for the task in the UI.

This diff specifically utilizes the above api to implement the rollback support for a failed switchover task. During swtichover a new xCluster config in the reverse direction is created and the old xCluster config is deleted. The rollback of a switchover is doable if the previous xCluster config is not yet deleted, otherwise, the user has to retry the failed task. This diff makes sure the right error is shown to the user in this case.

Test Plan:
- Aborted the switchover task in the middle and called the rollback api on the task and made sure the rollback task is created properly and the DR config end up in the right state after the rollback task finishes.

...TBD more tests...

Reviewers: #yba-api-review!, nsingh, cwang, jmak, yshchetinin, nbhatia

Reviewed By: cwang

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D40514