[#26126] YSQL: Plan hints enhancements.
Summary:
In their current form, plan hints have some deficiencies that make them difficult to use. Issues in the current implementation include:

  - Duplicate names/aliases in a query
  - Detection of hints with errors, or hints that lead to no valid plan being found
  - No corrective action taken currently
  - Hinting with multiple unnamed subqueries
  - Hinting with multiple Values clauses
  - Parallel hints do not work correctly in all cases

This feature includes improved hint error detection. Currently, only a small number of potential problems are handled:

  - Hint syntax errors
  - Duplicate table names in a single hint
  - Multiple join method hints for the same set of tables

However, several other problems can occur with other bad plan hints:

  - Bad index names, e.g. ‘Indexscan(t1 i1)’ if i1 does not exist
  - Hints that are not used, e.g., the hash join hint cannot be used since the join t1-t3 is prohibited by the Leading hint 'Leading(((t1 t2) t3)) Hashjoin(t1 t3)’
  - Hints that are unsatisfiable, i.e., that lead to no plan satisfying the hints being found. E.g., hash join has been specified but a nested loop join is required so a plan will be found but will have a ‘disabled’ (huge) cost : /*+ Hashjoin(t1 t2) */ SELECT 1 FROM t1, t2 WHERE a1<a2

The enhancements include new GUCs to help detect and fix these types of errors. The central one is pg_hint_plan.yb_bad_hint_mode, and it’s allowed values are:

  - OFF : Allow queries with “bad” hints to be planned even if hints have errors, are unused, or lead to undesired plans (default). A plan with a disabled cost may be found.
  - WARN : Issue a warning if any “bad” hints are found, and plan using hints where possible, i.e., some hints may get used and others ignored. A plan with a disabled cost may be found.
  - REPLAN : Same warnings issued for bad hints as with WARN but drops hints and replans (with no hints). A plan with a disabled cost will cause replanning.
  - ERROR : Same warnings issued as WARN but issues an error and no plan is produced. Any bad hint, or plan with disabled cost, will cause an error.

A major deficiency in the existing implementation is the lack of a means for hinting different blocks of a query that have tables with the same names or aliases. (A major goal of this work is to require no rewriting of original SQL.) To enable this, query-wide unique table aliases are generated and displayed in EXPLAIN output. The aliases displayed in the EXPLAIN may change from those assigned by Postgres, e.g. Result nodes do not normally have an alias displayed but with hinting unique names must be used to hint the plan.

To improve usability, and enhance the understanding of plan hints, a new EXPLAIN option has been added, namely a Boolean “HINTS”. If enabled, this shows the hints (with the label “Generated Hints”) that would force the exact same plan (in terms of shape and access/join methods) to be generated again.

pg_hint_plan includes a table called “hint_plan.hints” to store hints for a query. The format of this table (in the release currently used in Yugabyte) is:

     (id integer, norm_query_string text, application_name text , hints text)

Using the query text is awkward and error prone, and a much easier way to specify hints for a query is to use query id. To do this, a new GUC is provided, specifically pg_hint_plan.yb_use_query_id_for_hinting. If this is ON/TRUE, hints can be inserted in the hints table using query id as a text string instead of the query text.

A small change was made to the join search algorithm to safely prune joins that cannot possibly participate in the final plan. E.g., if one hints

    Leading(((t1 t2) t3))

then the 2-way joins t1-t3 and t2-t3 will be tried but will have ‘disable cost’ added to the cost of all of their paths (physical joins). If a valid plan is found for the join t1-t2 (as one would expect), then there is no reason to enumerate any further joins that start with the joins t1-t3 or t2-t3. This pruning greatly speeds up the planning process.

Another enhancement is the addition of planner tracing and unique node ids, which allows a user to understand the logic of the planner. This code will continue evolve.
Jira: DB-15453

Test Plan:
TestPgRegressExtension
TestPgRegressJoin
TestPgRegressAggregates
TestPgRegressYbBitmapScans
TestPgRegressMisc
TestPgRegressDistinctPushdown
TestPgRegressPlanner
TestPgRegressTAQO
TestPgRegressIndex
TestPgRegressPgIndex
TestPgRegressBackfillIndex
TestPgRegressThirdPartyExtensionsPgHintPlan
TestPgRegressPgSelect
TestPgRegressPartitions
TestPgRegressPgStat
TestPgRegressMatview
TestPgRegressPgMatview
TestPgExplainAnalyze
TestPgExplainAnalyzeJoins
TestPgRegressParallel
TestPgCardinalityEstimation
TestPgRegressContribPostgresFdw
TestPgRegressPgAuth
TestPgRegressPgMisc
TestPg15Regress
TestPgEstimatedDocdbResultWidth
TestPgRegressContribFileFdw
TestPgRegressCursor
TestPgRegressDml
TestPgRegressPgMiscIndependent
TestPgRegressPlpgsql
TestPgRegressPushdown
TestPgRegressThirdPartyExtensionsPgaudit
TestPgRegressWindow'
TestPgRegressYbIndexCheck
TestPgIndexSelectiveUpdate
TestPgRegressUpdateOptimized
TestPgRegressForeignKey

Reviewers: tnayak, gkukreja, mtakahara, mihnea, jason

Reviewed By: mtakahara

Subscribers: amartsinchyk, smishra, yql, svc_phabricator, jason

Differential Revision: https://phorge.dev.yugabyte.com/D41402