[PLAT-15172] Add warning banner to switchover modal when db versions are mismatched
Summary:
This diff add a warning banner to the YBA switchover modal which warns the user when they're
about to attempt switchover with mismatched database versions on DR primary and DR replica.
Switchover with mismatched database versions may run into issues with auto flag support so it
is best to run switchover only when the DR primary and DR replica are on the same database version.

Test Plan:
Create xCluster DR config between two universes on the same YBDB version.
Verify no warning banner appears in the switchover modal.
{F327032}
Upgrade one of the universes.
Verify that a warning banner appears in the switchover modal.
{F327031}

Reviewers: hzare, cwang, vbansal, rmadhavan

Reviewed By: rmadhavan

Subscribers: rmadhavan, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41389