[PLAT-16553]Restore universe on the same universe allowed to happen without renaming keyspaces
Summary:
In the Restore Modal, we previously restricted users from navigating to the next screen until the preflight check was completed. However, if the table API request was still pending, users were allowed to proceed to the next screen. This behavior inadvertently allowed users to bypass the mandatory "Rename Keyspace" screen.

Fix:

Now, we made the modal to disable the "next" button till the tables api is finished. And, also if the tables api is failed, we display an error message.

Test Plan:
Tested manually for both ysql and ycql
{F327588}

Reviewers: lsangappa

Reviewed By: lsangappa

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D41438