[#25382] YSQL: PG15 Online Upgrade: Delete previous version of the catalog tables after upgrade
Summary:
After the ysql major version upgrade completes, delete the previous version of the catalog tables since they are no longer needed.
The async task will keep retrying as long as the node is the master leader. We only expect to have transient failures, so this is safe.
If a new master leader is elected, and the cleanup is pending it will kick off a new task.

**Upgrade/Downgrade safety:**
New proto field `previous_version_catalog_cleanup_required` is added. This is set to non-default value only after the upgrade is finalized, so its safe.
Jira: DB-14612

Test Plan:
YsqlMajorUpgradeRpcsTest.SimultaneousUpgrades
YsqlMajorUpgradeRpcsTest.CleanupPreviousCatalog

Reviewers: telgersma, smishra

Reviewed By: telgersma

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D41420